+++ service_restaurant/src/main/kotlin/org/thechance/service_restaurant/api/usecases/CategoryUseCases.kt
@@ -2,6 +2,7 @@ package org.thechance.service_restaurant.api.usecases
 
 import org.koin.core.annotation.Single
 import org.thechance.service_restaurant.entity.Category
+import org.thechance.service_restaurant.entity.Restaurant
 
 
 @Single
@@ -10,7 +11,8 @@ data class CategoryUseCasesContainer(
     val getCategoryDetails: GetCategoryDetailsUseCase,
     val addCategory: CreateCategoryUseCases,
     val updateCategory: UpdateCategoryUseCases,
-    val deleteCategory: DeleteCategoryUseCases
+    val deleteCategory: DeleteCategoryUseCases,
+    val addRestaurantToCategory: AddRestaurantToCategoryUseCases
 )
 
 interface GetCategoriesUseCases {
@@ -31,4 +33,8 @@ interface UpdateCategoryUseCases {
 
 interface DeleteCategoryUseCases {
     suspend operator fun invoke(categoryId: String): Boolean
+}
+
+interface AddRestaurantToCategoryUseCases {
+    suspend operator fun invoke(categoryId: String, restaurantIds: List<String>): Boolean
 }
