+++ service_restaurant/src/main/kotlin/org/thechance/service_restaurant/api/usecases/RestaurantUseCases.kt
@@ -1,6 +1,16 @@
 package org.thechance.service_restaurant.api.usecases
 
+import org.koin.core.annotation.Single
 import org.thechance.service_restaurant.entity.Restaurant
+
+@Single
+data class RestaurantUseCasesContainer(
+    val getRestaurants: GetRestaurantsUseCase,
+    val getRestaurantDetails: GetRestaurantDetailsUseCase,
+    val addRestaurant: CreateRestaurantUseCase,
+    val updateRestaurant: UpdateRestaurantUseCase,
+    val deleteRestaurant: DeleteRestaurantUseCase
+)
 
 interface GetRestaurantsUseCase {
     suspend operator fun invoke(): List<Restaurant>
